+++ NEWS.md
@@ -1,4 +1,4 @@
-# nanonext 0.5.5.9008 (to be 0.6.0)
+# nanonext 0.5.5.9009 (to be 0.6.0)
 
 #### New Features
 
@@ -7,15 +7,15 @@
 
 #### Updates
 
-*Potentially breaking changes:*
+*Please review the following potentially breaking changes, and only update when ready:*
 
 * `send()` now returns an integer exit code in all cases. The 'echo' argument has been replaced by '...', and specifying 'echo' no longer has any effect.
+* `recv()`, `recv_aio()` and `request()` now set 'keep.raw' to FALSE as default to return only the originally sent object.
 * `ncurl()` argument 'request' renamed to 'response' for specifying response headers to return (to avoid confusion); the 'pem' argument is also moved after the new argument 'follow'.
 * `sha224()`, `sha256()`, `sha384()` and `sha512()` functions now take character vectors by serialising them rather than only taking the first element, whilst also being faster for character scalars. These functions also gain an argument 'convert' to control whether to return a raw vector or character string, instead of returning 'nanoHash' objects.
 
 *Other changes:*
 
-* `unresolved()` is faster and no longer forces resolution of the 'aio'. This allows the completion of an asynchronous operation to be checked without retrieving its value. For normal usage there is no behavioural difference, but in the special case where resolution is confirmed and the aio then stopped using `stop_aio()`, the values will not be available at `$raw` or `$data` as before.
 * `socket()` and `nano()` now accept non-missing NULL 'listen' and 'dial' arguments, allowing easier programmatic use.
 * Functions `send()`, `recv()`, `send_aio()`, `recv_aio()`, `setopt()`, `subscribe()`, `unsubscribe()` and `survey_time()` are no longer S3 generics for enhanced performance.
 * The source code of 'libnng' is now bundled instead of downloaded (more efficient as excludes unused portions).